chore: split searchbox implementation for samples in modal
